requirements.txt
streamlit>=1.28.0
google-ads>=24.0.0
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
google-auth-oauthlib>=1.0.0

